Restyled by black

+++ tensorflow/compiler/tests/special_math_test.py
@@ -37,9 +37,14 @@ from tensorflow.python.ops import gradient_checker_v2
 from tensorflow.python.ops import math_ops
 from tensorflow.python.platform import test
 
-flags.DEFINE_bool('vary_seed', False,
-                  ('Whether to vary the PRNG seed unpredictably.  '
-                   'With --runs_per_test=N, produces N iid runs.'))
+flags.DEFINE_bool(
+    "vary_seed",
+    False,
+    (
+        "Whether to vary the PRNG seed unpredictably.  "
+        "With --runs_per_test=N, produces N iid runs."
+    ),
+)
 
 NUM_SAMPLES = int(1e3)
 
@@ -56,47 +61,44 @@ def _igammac(a, x):
 
 # This is df/da / df/dx, where f = igamma.
 def implicit_reparameterization_grad(a, x):
-    log_prob = math_ops.xlogy(a - 1., x) - math_ops.lgamma(a) - x
+    log_prob = math_ops.xlogy(a - 1.0, x) - math_ops.lgamma(a) - x
     prob = math_ops.exp(log_prob)
     return -gen_math_ops.igamma_grad_a(a, x) / prob
 
 
 class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
-
     def setUp(self):
         if flags.FLAGS.vary_seed:
             entropy = os.urandom(64)
             if six.PY2:
-                answer = int(entropy.encode('hex'), 16)
+                answer = int(entropy.encode("hex"), 16)
             else:
-                answer = int.from_bytes(entropy, 'big')
-            np.random.seed(answer % (2**32 - 1))
+                answer = int.from_bytes(entropy, "big")
+            np.random.seed(answer % (2 ** 32 - 1))
         super(IgammaTest, self).setUp()
 
     # Skip Float64 test on TPU due to missing ops.
     def maybe_skip_test(self, dtype):
-        if self.device not in ['XLA_GPU', 'XLA_CPU'] and dtype == np.float64:
+        if self.device not in ["XLA_GPU", "XLA_CPU"] and dtype == np.float64:
             self.skipTest(
-                'Skipping test because some F64 operations not supported on TPU.')
+                "Skipping test because some F64 operations not supported on TPU."
+            )
 
     def adjust_tolerance_for_tpu(self, dtype, rtol, atol):
-        if self.device not in ['TPU']:
+        if self.device not in ["TPU"]:
             return rtol, atol
 
         if dtype == np.float32:
             return 2e-2, 1e-7
         return 2e-4, 1e-20
 
-    @parameterized.parameters((np.float32, 1e-2, 1e-11),
-                              (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 1e-2, 1e-11), (np.float64, 1e-4, 1e-30))
     def testLargeXSmallA(self, dtype, rtol, atol):
         self.maybe_skip_test(dtype)
         rtol, atol = self.adjust_tolerance_for_tpu(dtype, rtol, atol)
         # Test values near zero.
-        x = np.random.uniform(low=100., high=200., size=[
-                              NUM_SAMPLES]).astype(dtype)
-        a = np.random.uniform(low=0.3, high=1., size=[
-                              NUM_SAMPLES]).astype(dtype)
+        x = np.random.uniform(low=100.0, high=200.0, size=[NUM_SAMPLES]).astype(dtype)
+        a = np.random.uniform(low=0.3, high=1.0, size=[NUM_SAMPLES]).astype(dtype)
 
         expected_values = sps.gammainc(a, x)
         with self.session() as sess:
@@ -105,16 +107,17 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
             actual = sess.run(y)
         self.assertAllClose(expected_values, actual, atol=atol, rtol=rtol)
 
-    @parameterized.parameters((np.float32, 1e-2, 1e-11),
-                              (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 1e-2, 1e-11), (np.float64, 1e-4, 1e-30))
     def testSmallValues(self, dtype, rtol, atol):
         self.maybe_skip_test(dtype)
         rtol, atol = self.adjust_tolerance_for_tpu(dtype, rtol, atol)
         # Test values near zero.
         x = np.random.uniform(
-            low=np.finfo(dtype).tiny, high=1., size=[NUM_SAMPLES]).astype(dtype)
+            low=np.finfo(dtype).tiny, high=1.0, size=[NUM_SAMPLES]
+        ).astype(dtype)
         a = np.random.uniform(
-            low=np.finfo(dtype).tiny, high=1., size=[NUM_SAMPLES]).astype(dtype)
+            low=np.finfo(dtype).tiny, high=1.0, size=[NUM_SAMPLES]
+        ).astype(dtype)
 
         expected_values = sps.gammainc(a, x)
         with self.session() as sess:
@@ -122,16 +125,13 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
                 actual = sess.run(_igamma(a, x))
         self.assertAllClose(expected_values, actual, atol=atol, rtol=rtol)
 
-    @parameterized.parameters((np.float32, 1e-2, 1e-11),
-                              (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 1e-2, 1e-11), (np.float64, 1e-4, 1e-30))
     def testMediumValues(self, dtype, rtol, atol):
         self.maybe_skip_test(dtype)
         rtol, atol = self.adjust_tolerance_for_tpu(dtype, rtol, atol)
         # Test values near zero.
-        x = np.random.uniform(low=1., high=100., size=[
-                              NUM_SAMPLES]).astype(dtype)
-        a = np.random.uniform(low=1., high=100., size=[
-                              NUM_SAMPLES]).astype(dtype)
+        x = np.random.uniform(low=1.0, high=100.0, size=[NUM_SAMPLES]).astype(dtype)
+        a = np.random.uniform(low=1.0, high=100.0, size=[NUM_SAMPLES]).astype(dtype)
 
         expected_values = sps.gammainc(a, x)
         with self.session() as sess:
@@ -141,14 +141,16 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
 
     @parameterized.parameters((np.float32, 2e-2, 1e-5), (np.float64, 1e-4, 1e-30))
     def testLargeValues(self, dtype, rtol, atol):
-        if self.device == 'TPU':
+        if self.device == "TPU":
             # TODO(b/154908275): Remove this once fixed for large a, x.
-            self.skipTest('Skipping test since numerically unstable on TPU.')
+            self.skipTest("Skipping test since numerically unstable on TPU.")
         # Test values near zero.
-        x = np.random.uniform(
-            low=100., high=int(1e4), size=[NUM_SAMPLES]).astype(dtype)
-        a = np.random.uniform(
-            low=100., high=int(1e4), size=[NUM_SAMPLES]).astype(dtype)
+        x = np.random.uniform(low=100.0, high=int(1e4), size=[NUM_SAMPLES]).astype(
+            dtype
+        )
+        a = np.random.uniform(low=100.0, high=int(1e4), size=[NUM_SAMPLES]).astype(
+            dtype
+        )
 
         expected_values = sps.gammainc(a, x)
         with self.session() as sess:
@@ -164,15 +166,22 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
         with self.session():
             with self.test_scope():
                 x = constant_op.constant(
-                    np.random.uniform(low=1., high=100.,
-                                      size=[NUM_SAMPLES]).astype(dtype))
+                    np.random.uniform(low=1.0, high=100.0, size=[NUM_SAMPLES]).astype(
+                        dtype
+                    )
+                )
                 a = constant_op.constant(
-                    np.random.uniform(low=1., high=100.,
-                                      size=[NUM_SAMPLES]).astype(dtype))
+                    np.random.uniform(low=1.0, high=100.0, size=[NUM_SAMPLES]).astype(
+                        dtype
+                    )
+                )
 
-                def f(b): return _igamma(b, x)
+                def f(b):
+                    return _igamma(b, x)
+
                 max_error = gradient_checker_v2.max_error(
-                    *gradient_checker_v2.compute_gradient(f, x=[a], delta=1e-3))
+                    *gradient_checker_v2.compute_gradient(f, x=[a], delta=1e-3)
+                )
         self.assertLessEqual(max_error, tolerance)
 
     @parameterized.parameters((np.float32, 0.5), (np.float64, 1e-7))
@@ -181,19 +190,25 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
         with self.session():
             with self.test_scope():
                 x = constant_op.constant(
-                    np.random.uniform(low=100., high=int(1e4),
-                                      size=[NUM_SAMPLES]).astype(dtype))
+                    np.random.uniform(
+                        low=100.0, high=int(1e4), size=[NUM_SAMPLES]
+                    ).astype(dtype)
+                )
                 a = constant_op.constant(
-                    np.random.uniform(low=100., high=int(1e4),
-                                      size=[NUM_SAMPLES]).astype(dtype))
+                    np.random.uniform(
+                        low=100.0, high=int(1e4), size=[NUM_SAMPLES]
+                    ).astype(dtype)
+                )
 
-                def f(b): return _igamma(b, x)
+                def f(b):
+                    return _igamma(b, x)
+
                 max_error = gradient_checker_v2.max_error(
-                    *gradient_checker_v2.compute_gradient(f, x=[a], delta=1e-2))
+                    *gradient_checker_v2.compute_gradient(f, x=[a], delta=1e-2)
+                )
         self.assertLessEqual(max_error, tolerance)
 
-    @parameterized.parameters((np.float32, 1e-2, 1e-11),
-                              (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 1e-2, 1e-11), (np.float64, 1e-4, 1e-30))
     def testRandomGammaGradSmallValues(self, dtype, rtol, atol):
         self.maybe_skip_test(dtype)
         rtol, atol = self.adjust_tolerance_for_tpu(dtype, rtol, atol)
@@ -203,28 +218,31 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
             with self.test_scope():
                 x = constant_op.constant(
                     np.random.uniform(
-                        low=np.finfo(dtype).tiny, high=1.,
-                        size=[NUM_SAMPLES]).astype(dtype))
+                        low=np.finfo(dtype).tiny, high=1.0, size=[NUM_SAMPLES]
+                    ).astype(dtype)
+                )
                 a = constant_op.constant(
                     np.random.uniform(
-                        low=np.finfo(dtype).tiny, high=1.,
-                        size=[NUM_SAMPLES]).astype(dtype))
+                        low=np.finfo(dtype).tiny, high=1.0, size=[NUM_SAMPLES]
+                    ).astype(dtype)
+                )
                 gamma_sample_grad = gen_random_ops.random_gamma_grad(a, x)
                 actual_grad = implicit_reparameterization_grad(a, x)
                 gamma_sample_grad, actual_grad = sess.run(
-                    [gamma_sample_grad, actual_grad])
+                    [gamma_sample_grad, actual_grad]
+                )
                 # We do this because the ratio computed in
                 # implicit_reparameterization_grad can very easily result in a NaN due
                 # to the computed numerator and denominator zeroing out.
                 gamma_sample_grad = gamma_sample_grad[
-                    ~np.logical_or(np.isnan(actual_grad), np.isinf(actual_grad))]
+                    ~np.logical_or(np.isnan(actual_grad), np.isinf(actual_grad))
+                ]
                 actual_grad = actual_grad[
-                    ~np.logical_or(np.isnan(actual_grad), np.isinf(actual_grad))]
-        self.assertAllClose(actual_grad, gamma_sample_grad,
-                            atol=atol, rtol=rtol)
+                    ~np.logical_or(np.isnan(actual_grad), np.isinf(actual_grad))
+                ]
+        self.assertAllClose(actual_grad, gamma_sample_grad, atol=atol, rtol=rtol)
 
-    @parameterized.parameters((np.float32, 1e-2, 1e-11),
-                              (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 1e-2, 1e-11), (np.float64, 1e-4, 1e-30))
     def testRandomGammaGradMediumValues(self, dtype, rtol, atol):
         self.maybe_skip_test(dtype)
         rtol, atol = self.adjust_tolerance_for_tpu(dtype, rtol, atol)
@@ -232,63 +250,66 @@ class IgammaTest(xla_test.XLATestCase, parameterized.TestCase):
         with self.session() as sess:
             with self.test_scope():
                 x = constant_op.constant(
-                    np.random.uniform(low=1., high=10.,
-                                      size=[NUM_SAMPLES]).astype(dtype))
+                    np.random.uniform(low=1.0, high=10.0, size=[NUM_SAMPLES]).astype(
+                        dtype
+                    )
+                )
                 a = constant_op.constant(
-                    np.random.uniform(low=1., high=10.,
-                                      size=[NUM_SAMPLES]).astype(dtype))
+                    np.random.uniform(low=1.0, high=10.0, size=[NUM_SAMPLES]).astype(
+                        dtype
+                    )
+                )
                 gamma_sample_grad = gen_random_ops.random_gamma_grad(a, x)
                 actual_grad = implicit_reparameterization_grad(a, x)
                 gamma_sample_grad, actual_grad = sess.run(
-                    [gamma_sample_grad, actual_grad])
+                    [gamma_sample_grad, actual_grad]
+                )
                 # We do this because the ratio computed in
                 # implicit_reparameterization_grad can very easily result in a NaN due
                 # to the computed numerator and denominator zeroing out.
                 gamma_sample_grad = gamma_sample_grad[
-                    ~np.logical_or(np.isnan(actual_grad), np.isinf(actual_grad))]
+                    ~np.logical_or(np.isnan(actual_grad), np.isinf(actual_grad))
+                ]
                 actual_grad = actual_grad[
-                    ~np.logical_or(np.isnan(actual_grad), np.isinf(actual_grad))]
-        self.assertAllClose(actual_grad, gamma_sample_grad,
-                            atol=atol, rtol=rtol)
+                    ~np.logical_or(np.isnan(actual_grad), np.isinf(actual_grad))
+                ]
+        self.assertAllClose(actual_grad, gamma_sample_grad, atol=atol, rtol=rtol)
 
 
 class IgammacTest(xla_test.XLATestCase, parameterized.TestCase):
-
     def setUp(self):
         if flags.FLAGS.vary_seed:
             entropy = os.urandom(64)
             if six.PY2:
-                answer = int(entropy.encode('hex'), 16)
+                answer = int(entropy.encode("hex"), 16)
             else:
-                answer = int.from_bytes(entropy, 'big')
-            np.random.seed(answer % (2**32 - 1))
+                answer = int.from_bytes(entropy, "big")
+            np.random.seed(answer % (2 ** 32 - 1))
         super(IgammacTest, self).setUp()
 
     # Skip Float64 test on TPU due to missing ops.
     def maybe_skip_test(self, dtype):
-        if self.device not in ['XLA_GPU', 'XLA_CPU'] and dtype == np.float64:
+        if self.device not in ["XLA_GPU", "XLA_CPU"] and dtype == np.float64:
             # TODO(b/154908275): Remove this once fixed for large a, x.
             self.skipTest(
-                'Skipping test because some F64 operations not supported on TPU.')
+                "Skipping test because some F64 operations not supported on TPU."
+            )
 
     def adjust_tolerance_for_tpu(self, dtype, rtol, atol):
-        if self.device not in ['TPU']:
+        if self.device not in ["TPU"]:
             return rtol, atol
 
         if dtype == np.float32:
             return 2e-2, 1e-7
         return 2e-4, 1e-20
 
-    @parameterized.parameters((np.float32, 1e-2, 1e-11),
-                              (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 1e-2, 1e-11), (np.float64, 1e-4, 1e-30))
     def testLargeXSmallA(self, dtype, rtol, atol):
         self.maybe_skip_test(dtype)
         rtol, atol = self.adjust_tolerance_for_tpu(dtype, rtol, atol)
         # Test values near zero.
-        x = np.random.uniform(low=100., high=200., size=[
-                              NUM_SAMPLES]).astype(dtype)
-        a = np.random.uniform(low=0.3, high=1., size=[
-                              NUM_SAMPLES]).astype(dtype)
+        x = np.random.uniform(low=100.0, high=200.0, size=[NUM_SAMPLES]).astype(dtype)
+        a = np.random.uniform(low=0.3, high=1.0, size=[NUM_SAMPLES]).astype(dtype)
 
         expected_values = sps.gammaincc(a, x)
         with self.session() as sess:
@@ -297,16 +318,17 @@ class IgammacTest(xla_test.XLATestCase, parameterized.TestCase):
             actual = sess.run(y)
         self.assertAllClose(expected_values, actual, atol=atol, rtol=rtol)
 
-    @parameterized.parameters((np.float32, 1e-2, 1e-11),
-                              (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 1e-2, 1e-11), (np.float64, 1e-4, 1e-30))
     def testSmallValues(self, dtype, rtol, atol):
         self.maybe_skip_test(dtype)
         rtol, atol = self.adjust_tolerance_for_tpu(dtype, rtol, atol)
         # Test values near zero.
         x = np.random.uniform(
-            low=np.finfo(dtype).tiny, high=1., size=[NUM_SAMPLES]).astype(dtype)
+            low=np.finfo(dtype).tiny, high=1.0, size=[NUM_SAMPLES]
+        ).astype(dtype)
         a = np.random.uniform(
-            low=np.finfo(dtype).tiny, high=1., size=[NUM_SAMPLES]).astype(dtype)
+            low=np.finfo(dtype).tiny, high=1.0, size=[NUM_SAMPLES]
+        ).astype(dtype)
 
         expected_values = sps.gammaincc(a, x)
         with self.session() as sess:
@@ -314,16 +336,13 @@ class IgammacTest(xla_test.XLATestCase, parameterized.TestCase):
                 actual = sess.run(_igammac(a, x))
         self.assertAllClose(expected_values, actual, atol=atol, rtol=rtol)
 
-    @parameterized.parameters((np.float32, 1e-2, 1e-11),
-                              (np.float64, 1e-4, 1e-30))
+    @parameterized.parameters((np.float32, 1e-2, 1e-11), (np.float64, 1e-4, 1e-30))
     def testMediumValues(self, dtype, rtol, atol):
         self.maybe_skip_test(dtype)
         rtol, atol = self.adjust_tolerance_for_tpu(dtype, rtol, atol)
         # Test values near zero.
-        x = np.random.uniform(low=1., high=100., size=[
-                              NUM_SAMPLES]).astype(dtype)
-        a = np.random.uniform(low=1., high=100., size=[
-                              NUM_SAMPLES]).astype(dtype)
+        x = np.random.uniform(low=1.0, high=100.0, size=[NUM_SAMPLES]).astype(dtype)
+        a = np.random.uniform(low=1.0, high=100.0, size=[NUM_SAMPLES]).astype(dtype)
 
         expected_values = sps.gammaincc(a, x)
         with self.session() as sess:
@@ -333,13 +352,15 @@ class IgammacTest(xla_test.XLATestCase, parameterized.TestCase):
 
     @parameterized.parameters((np.float32, 2e-2, 1e-5), (np.float64, 1e-4, 1e-30))
     def testLargeValues(self, dtype, rtol, atol):
-        if self.device == 'TPU':
-            self.skipTest('Skipping test since numerically unstable on TPU.')
+        if self.device == "TPU":
+            self.skipTest("Skipping test since numerically unstable on TPU.")
         # Test values near zero.
-        x = np.random.uniform(
-            low=100., high=int(1e4), size=[NUM_SAMPLES]).astype(dtype)
-        a = np.random.uniform(
-            low=100., high=int(1e4), size=[NUM_SAMPLES]).astype(dtype)
+        x = np.random.uniform(low=100.0, high=int(1e4), size=[NUM_SAMPLES]).astype(
+            dtype
+        )
+        a = np.random.uniform(low=100.0, high=int(1e4), size=[NUM_SAMPLES]).astype(
+            dtype
+        )
 
         expected_values = sps.gammaincc(a, x)
         with self.session() as sess:
@@ -348,6 +369,6 @@ class IgammacTest(xla_test.XLATestCase, parameterized.TestCase):
         self.assertAllClose(expected_values, actual, atol=atol, rtol=rtol)
 
 
-if __name__ == '__main__':
-    os.environ['XLA_FLAGS'] = '--xla_cpu_enable_fast_math=false'
+if __name__ == "__main__":
+    os.environ["XLA_FLAGS"] = "--xla_cpu_enable_fast_math=false"
     test.main()
